Add support for current line highlight rendering.
This adds the VSCode line and gutter rendering options for the current
line as is available in VSCode.  These new settings are disabled by
default (even though line rendering is the VSCode default) and uses
the previous line highlighting format so as to not change behavior for
current users of the theme.

`highlight-thing' face has been added and it adapts its face based on
the setting of the line highlight rendering setting.